requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
pydantic
pyJWT 
passlib
python-multipart